+++ meraki_api/exceptions.py
@@ -1,0 +1,5 @@
+# meraki_ha/meraki_api/exceptions.py
+
+class MerakiApiError(Exception):
+    """Exception to indicate a Meraki API error."""
+    pass
